% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_character.R
\name{seeker_gen_pathway.character}
\alias{seeker_gen_pathway.character}
\title{Class character}
\usage{
\method{seeker_gen_pathway}{character}(x)
}
\arguments{
\item{x}{A gen symbol in character}
}
\value{
A data.frame with the pathways of the gen from reactome.org
}
\description{
seek_gen_pathway for the class [character]
}
\examples{
seeker_gen_pathway("MAPT")
}
