% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dataframe.R
\name{seeker_gen_pathway.data.frame}
\alias{seeker_gen_pathway.data.frame}
\title{Class data.frame}
\usage{
\method{seeker_gen_pathway}{data.frame}(x)
}
\arguments{
\item{x}{A genes symbols in dataframe}
}
\value{
A data.frame with the pathways of the gen from reactome.org
}
\description{
seek_gen_pathway for the class data.frame
}
\examples{
df <- data.frame(gen=c("MAPT", "APOE", "MMP12"))
seeker_gen_pathway(df)
}
