% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_default.R
\name{seeker_gen_pathway.default}
\alias{seeker_gen_pathway.default}
\title{Default method seeker_gen_pathway}
\usage{
\method{seeker_gen_pathway}{default}(x)
}
\arguments{
\item{x}{}
}
\value{
A message of error
}
\description{
This method return a error message when the function can use the class in the first parameter
}
\examples{
This method return an error message
}
