% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_factor.R
\name{seeker_gen_pathway.factor}
\alias{seeker_gen_pathway.factor}
\title{Class factor}
\usage{
\method{seeker_gen_pathway}{factor}(x)
}
\arguments{
\item{x}{A genes symbols in class factor}
}
\value{
A data.frame with the pathways of the gen from reactome.org
}
\description{
seek_gen_pathway for the class factor
}
\examples{
seeker_gen_pathway(factor("MAPT"))
}
