% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker_haploview.R
\name{seeker_haploview}
\alias{seeker_haploview}
\title{seeker_haploview}
\usage{
seeker_haploview(
  SNP,
  population_study = "1000GENOMES:phase_3:MXL",
  window_size = 500,
  d_prime = 0,
  color_select = "cyan"
)
}
\arguments{
\item{SNP}{A Single Nucleotide Polymorphism ID ("rs") in character.}

\item{population_study}{Population for which to compute LD. Default "1000GENOMES:phase_3:MXL"}
}
\value{
A correlation r2 plot between all SNPs with a Linkage Desequilibrium upper 0.5
}
\description{
A function that plot a correlation r2 between all SNPs with a Linkage Desequilibrium upper 0.5
}
\examples{
seeker_haploview("rs7412")

seeker_haploview("rs7412", population_study = "1000GENOMES:phase_3:FIN")

}
\author{
Erick Cuevas Fernández

Heriberto Manuel Rivera
}
