% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker_overlap.R
\name{seeker_overlap}
\alias{seeker_overlap}
\title{seeker overlap}
\source{
https://rest.ensembl.org
}
\usage{
seeker_overlap(id)
}
\arguments{
\item{id}{An Ensembl stable ID in string format}
}
\value{
A list with the features: Band, Gene, Transcript, CDS, Exon, Structural variation, Regulatory, Variation, Motif, ChipSeq.
}
\description{
A generic function to search features that overlap a region defined by the given identifier
}
\examples{
seeker_overlap("ENSG00000157764")


}
\author{
Erick Cuevas-Fernández

Heriberto Manuel Rivera
}
