% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker_snp_arch.R
\name{seeker_snp_arch}
\alias{seeker_snp_arch}
\alias{seeker_snp_arch.character}
\alias{seeker_snp_arch.data.frame}
\title{seeker Single Nucleotide Polymorphism (SNP) arquitecture}
\source{
https://rest.ensembl.org
}
\usage{
seeker_snp_arch(ID)

\method{seeker_snp_arch}{character}(ID)

\method{seeker_snp_arch}{data.frame}(ID)
}
\arguments{
\item{ID}{A Single Nucleotide Polymorphism (SNP) ID "rs" in character or data.frame}
}
\value{
A data.frame with the genomic arquitecture

\code{NULL}

\code{NULL}
}
\description{
seeker_snp_arch is a generic function to obtain the genome arquitecture data of a given SNP
}
\examples{
seeker_snp_arch("rs56116432")

df <- data.frame(c("rs56116432","rs10878307", "rs7133914", "rs11564148", "rs3761863", "rs10878245"))
seeker_snp_arch(df)

}
\author{
Erick Cuevas-Fernández
}
