% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker_snp_arq.R
\name{seeker_snp_arq}
\alias{seeker_snp_arq}
\title{seeker Single Nucleotide Polymorphism (SNP) arquitecture}
\source{
https://rest.ensembl.org
}
\usage{
seeker_snp_arq(ID)
}
\arguments{
\item{ID}{A Single Nucleotide Polymorphism (SNP) ID "rs" in character or data.frame}
}
\value{
A data.frame with the genomic arquitecture
}
\description{
seeker_snp_arq is a generic function to obtain the genome arquitecture data of a given SNP
}
\examples{
seeker_snp_arq("rs56116432")

df <- data.frame(c("rs56116432","rs10878307", "rs7133914", "rs11564148", "rs3761863", "rs10878245"))
seeker_snp_arq(df)

}
\author{
Erick Cuevas-Fernández

Heriberto Manuel Rivera
}
