% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker_snp_context.R
\name{seeker_snp_context}
\alias{seeker_snp_context}
\alias{seeker_snp_context.factor}
\alias{seeker_snp_context.character}
\alias{seeker_snp_context.data.frame}
\title{seeker Single Nucleotide Polymorphism Context}
\source{
https://rest.ensembl.org
}
\usage{
seeker_snp_context(SNP)

\method{seeker_snp_context}{factor}(SNP)

\method{seeker_snp_context}{character}(SNP)

\method{seeker_snp_context}{data.frame}(SNP)
}
\arguments{
\item{SNP}{A Single Nucleotide Polymorphism ID ("rs") in character or data.frame}
}
\value{
A data.frame with Type of variation
}
\description{
A generic function to search the type of variation & gene
}
\examples{
seeker_snp_context("rs56116432")

df <- data.frame(c("rs56116432","rs10878307", "rs7133914", "rs11564148", "rs3761863", "rs10878245"))
seeker_snp_context(df)

}
\author{
Erick Cuevas-Fernández

Heriberto Manuel Rivera
}
