% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker_snp_freq.R
\name{seeker_snp_freq}
\alias{seeker_snp_freq}
\alias{seeker_snp_freq.character}
\alias{seeker_snp_freq.data.frame}
\title{seeker Single Nucleotide Polymorphism frequency}
\source{
https://rest.ensembl.org
}
\usage{
seeker_snp_freq(ID, study = "1000GENOMES:phase_3")

\method{seeker_snp_freq}{character}(ID, study = "1000GENOMES:phase_3")

\method{seeker_snp_freq}{data.frame}(ID, study = "1000GENOMES:phase_3")
}
\arguments{
\item{ID}{A Single Nucleotide Polymorphism ID ("rs") in character or data.frame}

\item{study}{A study of population frequency. Default ("1000GENOMES:phase3")}
}
\value{
A data.frame with the allel frequency in all the population of the given study

\code{NULL}

\code{NULL}
}
\description{
A generic function to search the population frequency of a specific study
}
\examples{
seeker_snp_freq("rs56116432")

df <- data.frame(c("rs56116432","rs10878307", "rs7133914", "rs11564148", "rs3761863", "rs10878245"))
seeker_snp_freq(df)

}
