% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker_snp_freq_format.R
\name{seeker_snp_freq_format}
\alias{seeker_snp_freq_format}
\title{seeker_snp_freq_format}
\usage{
seeker_snp_freq_format(data)
}
\arguments{
\item{data}{A data.frame, output of seeker_snp_freq()}
}
\value{
A data.frame with the allel frequency in a horizontal format, with sorted frequencies.
}
\description{
seeker_snp_freq_format
}
\examples{

df <- seeker_snp_freq("rs56116432")

seeker_snp_freq_format(df)

}
\author{
Erick Cuevas-Fernández
}
