% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker_snp_ld.R
\name{seeker_snp_ld}
\alias{seeker_snp_ld}
\alias{seeker_snp_ld.character}
\alias{seeker_snp_ld.data.frame}
\title{seeker Single Nucleotide Polymorphism Linkage-Desequilibrium}
\source{
https://rest.ensembl.org
}
\usage{
seeker_snp_ld(ID, population = "1000GENOMES:phase_3:MXL",
  window_size = 500, d_prime = 0)

\method{seeker_snp_ld}{character}(ID,
  population = "1000GENOMES:phase_3:MXL", window_size = 500,
  d_prime = 0)

\method{seeker_snp_ld}{data.frame}(ID,
  population = "1000GENOMES:phase_3:MXL", window_size = 500,
  d_prime = 0)
}
\arguments{
\item{ID}{A Single Nucleotide Polymorphism ID ("rs") in character or data.frame}

\item{population}{Population for which to compute LD. Default "1000GENOMES:phase_3:MXL"}

\item{window_size}{Window size in kb. The maximum allowed value for the window size is 500 kb.
LD is computed for the given variant and all variants that are located within the specified window.
Default 500}

\item{d_prime}{Measure of LD. If D' is provided only return pairs of variants whose D' value is equal to or greater than the value provided.
Default 0}
}
\value{
A data.frame with the LD information of SNP in a specific population

\code{NULL}

\code{NULL}
}
\description{
seeker Single Nucleotide Polymorphism Linkage-Desequilibrium
}
\examples{
seeker_snp_ld("rs56116432")

df <- data.frame(c("rs56116432","rs10878307", "rs7133914", "rs11564148", "rs3761863", "rs10878245"))
seeker_snp_ld(df)

}
\author{
Erick Cuevas-Fernández
}
