% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker_snp_ld_plot.R
\name{seeker_snp_ld_plot}
\alias{seeker_snp_ld_plot}
\title{seeker_snp_ld_plot}
\usage{
seeker_snp_ld_plot(SNP, population_study = "1000GENOMES:phase_3:MXL",
  window_size = 500, d_prime = 0, color_select = "cyan")
}
\arguments{
\item{SNP}{A Single Nucleotide Polymorphism ID ("rs") in character.}

\item{population_study}{Population for which to compute LD. Default "1000GENOMES:phase_3:MXL"}

\item{window_size}{Window size in kb. The maximum allowed value for the window size is 500 kb.
LD is computed for the given variant and all variants that are located within the specified window.
Default 500}

\item{d_prime}{Measure of LD. If D' is provided only return pairs of variants whose D' value is equal to or greater than the value provided.
Default 0}

\item{color_select}{Select a color for the plot; Default= "cyan"}
}
\value{
A plot, in x axis the position in the chromosome, in y axis the r2 of linkage desequilibrium
}
\description{
A function that plot the Linkage-Dsequilibrium information for a given Single Nucleotide Polymorphism
}
\examples{
seeker_snp_ld_plot("rs7412", "1000GENOMES:phase_3:MXL", color_select = "green")

seeker_snp_ld_plot("rs7412", "1000GENOMES:phase_3:SAS", color_select = "pink")

}
\author{
Erick Cuevas Fernández

Heriberto Manuel Rivera
}
